requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
drf-spectacular>=0.27
pandas>=2.0
openpyxl>=3.1
pytest>=8.0
pytest-django>=4.8